Move page to contacts dir
